Refactor tests
1. REST APIs were already removed, so these are not needed.
2. Define common conftest fixtures and test factories  in the
   main project folder and then import them in other apps.
3. Rename context_value to context in query calls to GraphQL.
